Add Eval ('eval') and EvalInPlace('inplace').
Eval evaluates a string as a separate program within the current stack.
EvalInPlace evaluates a string as if it were pasted at the call site.

Impl Eval command: evaluates a string as a program and returns result

Add EvalInPlace command. Simluates copy-and-pasting argument string at call site and executing as normal

Manually implement match for ExecSignals